move new stream to Utility
