+++ Duplicati/Library/Main/Operation/Backup/StreamBlock.cs
@@ -14,10 +14,9 @@
 //  You should have received a copy of the GNU Lesser General Public
 //  License along with this library; if not, write to the Free Software
 //  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
-using System;
+
 using CoCoL;
 using System.IO;
-using System.Threading;
 using System.Threading.Tasks;
 using Duplicati.Library.Interface;
 
@@ -43,7 +42,7 @@ namespace Duplicati.Library.Main.Operation.Backup
             var tcs = new TaskCompletionSource<StreamProcessResult>();
 
             // limit the stream length to that found now, a fixed point in time
-            var limitedStream = new StreamReadLimitLengthWrapper(stream, stream.Length);
+            var limitedStream = new Library.Utility.ReadLimitLengthStream(stream, stream.Length);
             
             var streamBlock = new StreamBlock
             {
@@ -60,146 +59,4 @@ namespace Duplicati.Library.Main.Operation.Backup
         }
     }
 
-    // StreamReadLimitLengthWrapper() based on code from Matt Smith (Oct 26 '15 at 20:38)
-    // https://stackoverflow.com/questions/33354822/how-to-set-length-in-stream-without-truncated
-    sealed class StreamReadLimitLengthWrapper : Stream
-    {
-        readonly Stream m_innerStream;
-        readonly long m_endPosition;
-
-        public StreamReadLimitLengthWrapper(Stream innerStream, long size)
-        {
-            if (size < 0)
-            {
-                throw new ArgumentOutOfRangeException(nameof(size));
-            }
-
-            m_innerStream = innerStream ?? throw new ArgumentNullException(nameof(innerStream));
-            m_endPosition = m_innerStream.Position + size;
-        }
-
-        public override bool CanRead => m_innerStream.CanRead;
-
-        public override bool CanSeek => m_innerStream.CanSeek;
-
-        public override bool CanWrite => false;
-
-        public override void Flush()
-        {
-            m_innerStream.Flush();
-        }
-
-        public override long Length => m_endPosition;
-
-        public override long Position
-        {
-            get => m_innerStream.Position;
-            set => m_innerStream.Position = value;
-        }
-
-        public override int Read(byte[] buffer, int offset, int count)
-        {
-            count = GetAllowedCount(count);
-            return m_innerStream.Read(buffer, offset, count);
-        }
-
-        public override long Seek(long offset, SeekOrigin origin)
-        {
-            return m_innerStream.Seek(offset, origin);
-        }
-
-        public override void SetLength(long value)
-        {
-            throw new NotSupportedException();
-        }
-
-        public override void Write(byte[] buffer, int offset, int count)
-        {
-            throw new NotSupportedException();
-        }
-
-        public override bool CanTimeout => m_innerStream.CanTimeout;
-
-        public override int ReadTimeout
-        {
-            get => m_innerStream.ReadTimeout;
-            set => m_innerStream.ReadTimeout = value;
-        }
-
-        public override int WriteTimeout
-        {
-            get => m_innerStream.ReadTimeout;
-            set => m_innerStream.ReadTimeout = value;
-        }
-
-        public override IAsyncResult BeginRead(byte[] buffer, int offset, int count, AsyncCallback callback, object state)
-        {
-            count = GetAllowedCount(count);
-            return m_innerStream.BeginRead(buffer, offset, count, callback, state);
-        }
-
-        public override IAsyncResult BeginWrite(byte[] buffer, int offset, int count, AsyncCallback callback, object state)
-        {
-            throw new NotSupportedException();
-        }
-
-        public override void Close()
-        {
-            // Since this wrapper does not own the underlying stream, we do not want it to close the underlying stream
-        }
-
-        public override Task CopyToAsync(Stream destination, int bufferSize, CancellationToken cancellationToken)
-        {
-            return m_innerStream.CopyToAsync(destination, bufferSize, cancellationToken);
-        }
-
-        public override int EndRead(IAsyncResult asyncResult)
-        {
-            return m_innerStream.EndRead(asyncResult);
-        }
-
-        public override Task FlushAsync(CancellationToken cancellationToken)
-        {
-            return m_innerStream.FlushAsync(cancellationToken);
-        }
-
-        public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-        {
-            count = GetAllowedCount(count);
-            return m_innerStream.ReadAsync(buffer, offset, count, cancellationToken);
-        }
-
-        public override int ReadByte()
-        {
-            var count = GetAllowedCount(1);
-            if (count == 0)
-            {
-                return -1;
-            }
-
-            return m_innerStream.ReadByte();
-        }
-
-        public override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-        {
-            throw new NotSupportedException();
-        }
-
-        public override void WriteByte(byte value)
-        {
-            throw new NotSupportedException();
-        }
-
-        private int GetAllowedCount(int count)
-        {
-            long pos = m_innerStream.Position;
-            long maxCount = m_endPosition - pos;
-            if (count > maxCount)
-            {
-                return (int)maxCount;
-            }
-
-            return count;
-        }
-    }
 }
