+++ Duplicati/Library/Utility/Duplicati.Library.Utility.csproj
@@ -56,6 +56,7 @@
     <Compile Include="OverrideableStream.cs" />
     <Compile Include="ProgressReportingStream.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ReadLimitLengthStream.cs" />
     <Compile Include="ShaderStream.cs" />
     <Compile Include="Sizeparser.cs" />
     <Compile Include="SslCertificateValidator.cs" />
